chore: new trait methods
