+++ rust/main/submitter/Cargo.toml
@@ -15,6 +15,7 @@ async-trait.workspace = true
 chrono.workspace = true
 derive-new.workspace = true
 eyre.workspace = true
+futures-util.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
