+++ rust/main/submitter/src/payload_dispatcher.rs
@@ -1,11 +1,10 @@
 // TODO: re-enable clippy warnings
 #![allow(unused_imports)]
 
-mod building_stage;
 mod dispatcher;
 mod entrypoint;
-mod finality_stage;
-mod inclusion_stage;
+
+mod stages;
 mod test_utils;
 
 pub use dispatcher::*;
